Refactor Auth and UserController

+++ src/main/java/edu/ukma/projectmanagementsystem/service/security/AuthService.java
@@ -1,36 +1,33 @@
 package edu.ukma.projectmanagementsystem.service.security;
 
+import edu.ukma.projectmanagementsystem.config.AuthenticationFacade;
 import edu.ukma.projectmanagementsystem.config.TokenData;
 import edu.ukma.projectmanagementsystem.domain.entity.RefreshTokenEntity;
-import edu.ukma.projectmanagementsystem.domain.entity.UserEntity;
-import edu.ukma.projectmanagementsystem.domain.enumerated.UserRole;
-import edu.ukma.projectmanagementsystem.domain.repository.UserRepository;
-import edu.ukma.projectmanagementsystem.service.security.dto.JwtResponseDto;
-import edu.ukma.projectmanagementsystem.service.security.dto.LoginDto;
-import edu.ukma.projectmanagementsystem.service.security.dto.RefreshTokenRequest;
+import edu.ukma.projectmanagementsystem.service.business.UserService;
+import edu.ukma.projectmanagementsystem.service.dto.security.JwtResponseDto;
+import edu.ukma.projectmanagementsystem.service.dto.security.LoginDto;
+import edu.ukma.projectmanagementsystem.service.dto.security.RefreshTokenRequest;
+import edu.ukma.projectmanagementsystem.service.dto.user.UserDto;
 import lombok.RequiredArgsConstructor;
 import org.springframework.http.HttpStatus;
-import org.springframework.security.authentication.BadCredentialsException;
+import org.springframework.security.authentication.AnonymousAuthenticationToken;
 import org.springframework.security.core.Authentication;
-import org.springframework.security.crypto.password.PasswordEncoder;
 import org.springframework.stereotype.Service;
 import org.springframework.web.server.ResponseStatusException;
 
 @Service
 @RequiredArgsConstructor
 public class AuthService {
-    private final PasswordEncoder passwordEncoder;
-    private final UserRepository userRepository;
+    private final AuthenticationFacade authenticationFacade;
     private final RefreshTokenService refreshTokenService;
     private final JwtService jwtService;
+    private final UsernamePasswordAuthenticator usernamePasswordAuthenticator;
+    private final UserService userService;
 
     public JwtResponseDto authenticate(LoginDto loginDto) {
-        UserEntity user = userRepository.findByEmail(loginDto.getEmail())
-                .orElseThrow(() -> new BadCredentialsException("Invalid email or password"));
-        if (!passwordEncoder.matches(loginDto.getPassword(), user.getPassword())) {
-            throw new BadCredentialsException("Invalid email or password");
-        }
-        String token = jwtService.generateToken(new TokenData(user.getId(), UserRole.valueOf(user.getRole().getName())));
+        usernamePasswordAuthenticator.authenticateUser(loginDto);
+        UserDto user = userService.findUserByEmail(loginDto.getEmail());
+        String token = jwtService.generateToken(new TokenData(user.getId(), user.getRole()));
         RefreshTokenEntity refreshToken = refreshTokenService.createRefreshToken(user.getId());
         return new JwtResponseDto(user.getId(), token, refreshToken.getToken());
     }
@@ -41,13 +38,16 @@ public class AuthService {
                 .map(refreshTokenService::verifyExpiration)
                 .map(RefreshTokenEntity::getUserId)
                 .orElseThrow(() -> new ResponseStatusException(HttpStatus.UNAUTHORIZED));
-        UserEntity user = userRepository.findById(userId).orElseThrow(() -> new ResponseStatusException(HttpStatus.FORBIDDEN));
-        String token = jwtService.generateToken(new TokenData(user.getId(), UserRole.valueOf(user.getRole().getName())));
+        UserDto user = userService.findUserById(userId);
+        String token = jwtService.generateToken(new TokenData(user.getId(), user.getRole()));
         return new JwtResponseDto(userId, token, requestRefreshToken);
     }
 
-    public void logout(Authentication authentication) {
-        TokenData tokenData = (TokenData) authentication.getPrincipal();
-        refreshTokenService.deleteByUserId(tokenData.getId());
+    public void logoutCurrentUser() {
+        Authentication authentication = authenticationFacade.getAuthentication();
+        if (!(authentication instanceof AnonymousAuthenticationToken)){
+            TokenData tokenData = (TokenData) authentication.getPrincipal();
+            refreshTokenService.deleteByUserId(tokenData.getId());
+        }
     }
 }
